requirements.used.txt
accelerate==1.12.0
addict==2.4.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
backports.tarfile==1.2.0
bitsandbytes==0.45.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.36.0
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
filelock==3.17.0
fsspec==2025.2.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
groundingdino==0.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
invisible-watermark==0.2.0
jaraco.context==5.3.0
Jinja2==3.1.5
joblib==1.5.0
jupyter_core==5.9.1
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
more-itertools==10.3.0
mpmath==1.3.0
numpy==1.26.4
open_clip_torch==2.30.0
openai==1.78.0
opencv-python==4.11.0.86
optimum-quanto==0.2.4
orjson==3.10.15
packaging==25.0
pandas==2.2.3
peft==0.17.0
pillow==11.1.0
platformdirs==4.3.6
protobuf==3.20.3
psutil==6.1.1
pyarrow==19.0.1
pycocotools==2.0.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
sentry-sdk==2.21.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
supervision==0.26.0rc7
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.14
tokenizers==0.22.2
torch==2.5.1+cu121
torchmetrics==0.5.0
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.6
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.6
wcwidth==0.2.14
yapf==0.43.0
git+https://github.com/facebookresearch/segment-anything.git